requirements.txt
python-binance==1.0.19
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
